chore: move tests
